Fixed Difficulty menu, top black bar, and random events script
difficulty menu now works when you start a game you can choose the stats you want to start with. Top black bar not disappearing when going into policy menu fixed. Tried to fix error in random events script where it was making an object of a random event but that was giving errors.